hardware information
Added files for the schematics and layouts of the PDQ boards. They
include a correction for the amplifier IC, which is the AD8250 (see BoM
for all the correct parts). This also includes images for the by-hand
fix to solve clock ringing on a master board due to impedence
mismatches.